chore(ipv4): all logic into ipv4.rs
